# packages/contracts/openapi/organization.yaml
openapi: 3.1.0
info:
  title: 联软CRM V3组织架构与企微目录同步接口契约
  version: 2026-08-26
  description: |
    本契约描述当前已挂载的组织架构、角色授权和企业微信目录同步接口。
    所有接口使用同源签名 Cookie 会话，且当前实现只允许 superadmin 访问；
    不接受 operatorId、用户名、兼容令牌或其他客户端自报身份作为授权依据。
    组织功能由 V3_ORGANIZATION_ENABLED 控制，写入还受 V3_ORGANIZATION_WRITE_ENABLED 控制；
    账号入口整合另由 V3_ORGANIZATION_ACCOUNT_ENTRY_MERGED 控制，且仅影响管理端导航呈现；
    停用账号旧会话防护受 V3_AUTH_ACCOUNT_STATUS_CHECK_ENABLED 独立控制；
    停用归档交接额外受 V3_ORGANIZATION_OFFBOARDING_ENABLED 控制，且开启前必须先开启账号状态防护；
    企微目录同步额外受 V3_DIRECTORY_SYNC_ENABLED 控制。以上开关默认均为 false。
    企微目录同步首期仅创建和读取 integration 域预览数据，绝不改写组织、账号、渠道或 CRM 主体数据。
    企业微信身份映射导入仅写入 iam.external_identities，必须由超级管理员上传、预览并确认；不改变企微目录同步、泛微 OA、订单或自动拉群规则。
servers:
  - url: /
    description: V3 同源服务入口
security:
  - V3签名会话: []
...
paths:
...
  /api/rbac/accounts:
    get:
      tags: [角色与权限]
      operationId: 检索组织账号
      summary: 按用户名或姓名检索账号
      description: 仅超级管理员可调用。返回账号、系统角色以及内部任职和渠道成员关系摘要；不返回密码散列、外部身份主体或扩展原始数据。
      parameters:
        - name: keyword
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "200": { $ref: "#/components/responses/列表成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
  /api/rbac/roles/{id}/users:
    get:
      tags: [角色与权限]
      operationId: 查询角色关联用户
      summary: 分页查询拥有指定系统角色的用户
      description: 仅超级管理员可调用。用户详情仅返回用于定位账号和核对组织归属的最小必要字段。
      parameters:
        - $ref: "#/components/parameters/组织编号"
        - name: keyword
          in: query
          schema: { type: string, maxLength: 100 }
        - name: includeInactive
          in: query
          schema: { type: boolean, default: false }
          description: 默认仅返回启用账号；为 true 时追加停用历史账号。
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200": { $ref: "#/components/responses/分页成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
        "404": { $ref: "#/components/responses/组织不存在" }
...
components:
  securitySchemes:
    V3签名会话:
      type: apiKey
      in: cookie
      name: v3_session
      description: V3 HttpOnly 签名会话 Cookie；具体 Cookie 名称由部署配置维护，客户端不得伪造。
  parameters:
...
    组织编号:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
  responses:
...
    列表成功:
      description: 列表读取成功。
      content: { application/json: { schema: { $ref: "#/components/schemas/列表成功响应" } } }
    分页成功:
      description: 分页读取成功。
      content: { application/json: { schema: { $ref: "#/components/schemas/分页成功响应" } } }
...
    组织未登录:
      description: ORG_AUTH_REQUIRED、V3_AUTH_ACCOUNT_DISABLED、V3_AUTH_ACCOUNT_UNAVAILABLE 或 V3_AUTH_SESSION_MISMATCH
      content: { application/json: { schema: { $ref: "#/components/schemas/错误响应" } } }
    组织无权限:
      description: ORG_PERMISSION_DENIED
      content: { application/json: { schema: { $ref: "#/components/schemas/错误响应" } } }
...
    组织请求不合法:
      description: ORG_REQUEST_INVALID 或 ORG_IDEMPOTENCY_KEY_INVALID
      content: { application/json: { schema: { $ref: "#/components/schemas/错误响应" } } }
...
    组织不存在:
      description: ORG_NOT_FOUND
      content: { application/json: { schema: { $ref: "#/components/schemas/错误响应" } } }
...
  schemas:
    UUID: { type: string, format: uuid }
    通用成功响应:
      type: object
      required: [success, data, meta]
      properties:
        success: { const: true }
        data: { type: object, additionalProperties: true }
        meta:
          type: object
          required: [requestId, build]
          properties:
            requestId: { type: string }
            build: { type: object, additionalProperties: true }
    列表成功响应:
      allOf:
        - $ref: "#/components/schemas/通用成功响应"
        - type: object
          properties:
            data:
              type: object
              required: [items]
              properties:
                { items: { type: array, items: { type: object, additionalProperties: true } } }
    分页成功响应:
      allOf:
        - $ref: "#/components/schemas/通用成功响应"
        - type: object
          properties:
            data:
              type: object
              required: [items, page, pageSize, total]
              properties:
                items: { type: array, items: { type: object, additionalProperties: true } }
                page: { type: integer, minimum: 1 }
                pageSize: { type: integer, minimum: 1 }
                total: { type: integer, minimum: 0 }
...
    错误响应:
      type: object
      required: [success, error, meta]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string, description: 中文可读错误消息。 }
        meta:
          type: object
          required: [requestId, build]
          properties:
            requestId: { type: string }
            build: { type: object, additionalProperties: true }
...